feat: add version field to bom, change price to float, add migration for bom versions table

+++ src/domain/models/component.rs
@@ -1,13 +1,13 @@
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub struct Price {
-    pub value: i32,
+    pub value: f32,
     pub currency: String,
 }
 
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub struct Component {
     pub id: Uuid,
     pub name: String,
